refactor: 나의 카테고리 목록의 요소 컴포넌트 구현 및 분리
